Extract model and parse
