+++ model.go
@@ -1,0 +1,6 @@
+package main
+
+type Model struct {
+	Name   string
+	Fields map[string]string
+}
